Fix lib issues and restructure project
