+++ priv/templates/nif.ex.eex
@@ -1,21 +1,21 @@
 defmodule <%= mod %> do
-    # When your move/rename this file (and you should) make sure
-    # you also change the ERL_NIF_INIT call at the bottom of
-    # c_src/#{name}_nif.c
-    @on_load :init
+  # When your move/rename this file (and you should) make sure
+  # you also change the ERL_NIF_INIT call at the bottom of
+  # c_src/#{name}_nif.c
+  @on_load :init
 
-    def init do
-      :erlang.load_nif("./priv/lib/lib_<%= name %>", 0)
-    end
+  def init do
+    path = :filename.join(:code.priv_dir(:<%= name %>), 'lib_<%= name %>')
+    :ok  = :erlang.load_nif(path, 1)
+  end
 
-    # A simple wrapper around the NIF call
-    def hello do
-      _hello
-    end
+  # A simple wrapper around the NIF call
+  def hello do
+    _hello
+  end
 
-    # This function will not be overwritten if the nif fails to load
-    def _hello do
-      "NIF library not loaded"
-    end
-
+  # This function will not be overwritten if the nif fails to load
+  def _hello do
+    exit(:nif_library_not_loaded)
+  end
 end
